+++ stream-chat-android-client/src/main/java/io/getstream/chat/android/client/api2/model/dto/MemberDtos.kt
@@ -73,3 +73,13 @@ internal data class DownstreamMemberDto(
     val archived_at: Date?,
     val extraData: Map<String, Any>,
 ) : ExtraDataDto
+
+/**
+ * DTO holding limited data about a channel member.
+ *
+ * @property channel_role The role of the member in the channel.
+ */
+@JsonClass(generateAdapter = true)
+internal data class DownstreamMemberInfoDto(
+    val channel_role: String?,
+)