+++ stream-chat-android-client/src/test/java/io/getstream/chat/android/client/parser2/testdata/MemberInfoDtoTestData.kt
@@ -1,0 +1,32 @@
+/*
+ * Copyright (c) 2014-2025 Stream.io Inc. All rights reserved.
+ *
+ * Licensed under the Stream License;
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    https://github.com/GetStream/stream-chat-android/blob/main/LICENSE
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package io.getstream.chat.android.client.parser2.testdata
+
+import io.getstream.chat.android.client.api2.model.dto.DownstreamMemberInfoDto
+import org.intellij.lang.annotations.Language
+
+internal object MemberInfoDtoTestData {
+
+    @Language("JSON")
+    val downstreamJson: String = """
+        {
+          "channel_role": "channel_member"
+        }
+    """.trimIndent()
+
+    val downstreamMemberInfo = DownstreamMemberInfoDto(channel_role = "channel_member")
+}