+++ stream-chat-android-compose/src/main/java/io/getstream/chat/android/compose/ui/components/messages/MessageReactions.kt
@@ -20,17 +20,21 @@ import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.RowScope
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.platform.testTag
+import androidx.compose.ui.res.pluralStringResource
+import androidx.compose.ui.semantics.contentDescription
+import androidx.compose.ui.semantics.semantics
+import androidx.compose.ui.semantics.testTag
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
+import io.getstream.chat.android.compose.R
 import io.getstream.chat.android.compose.previewdata.PreviewReactionOptionData
 import io.getstream.chat.android.compose.state.reactionoptions.ReactionOptionItemState
 import io.getstream.chat.android.compose.ui.theme.ChatTheme
+import io.getstream.chat.android.compose.ui.theme.MessageReactionItemParams
 
 /**
  * Represents a reaction bubble with a list of reactions this message has.
@@ -43,22 +47,29 @@ import io.getstream.chat.android.compose.ui.theme.ChatTheme
 public fun MessageReactions(
     options: List<ReactionOptionItemState>,
     modifier: Modifier = Modifier,
-    itemContent: @Composable RowScope.(ReactionOptionItemState) -> Unit = { option ->
-        MessageReactionItem(
-            modifier = Modifier
-                .size(20.dp)
-                .padding(2.dp)
-                .align(Alignment.CenterVertically)
-                .testTag("Stream_MessageReaction_${option.type}"),
-            option = option,
-        )
+    itemContent: @Composable RowScope.(ReactionOptionItemState) -> Unit = { state ->
+        with(ChatTheme.componentFactory) {
+            MessageReactionItem(
+                params = MessageReactionItemParams(
+                    state = state,
+                ),
+            )
+        }
     },
 ) {
+    val description = pluralStringResource(
+        R.plurals.stream_ui_message_list_semantics_message_reactions,
+        options.size,
+        options.size,
+    )
     Row(
         modifier = modifier
+            .semantics {
+                testTag = "Stream_MessageReaction"
+                contentDescription = description
+            }
             .background(shape = RoundedCornerShape(16.dp), color = ChatTheme.colors.barsBackground)
-            .padding(4.dp)
-            .testTag("Stream_MessageReaction"),
+            .padding(4.dp),
         verticalAlignment = Alignment.CenterVertically,
     ) {
         options.forEach { option ->