+++ CHANGELOG.md
@@ -14,6 +14,10 @@
 
 ### ❌ Removed
 
+## stream-chat-android-core
+### ✅ Added
+- Add `Channel.messageCount` field representing the total number of messages in the channel. [#5918](https://github.com/GetStream/stream-chat-android/pull/5918)
+
 ## stream-chat-android-client
 ### 🐞 Fixed
 - Fix API_KEY being sent with the request done to the CDN when sharing a video from the media gallery. [#5921](https://github.com/GetStream/stream-chat-android/pull/5921)
@@ -23,6 +27,9 @@
 ### ✅ Added
 - Add `ChatClient.getPendingMessage` to fetch a pending message (and its metadata) by its ID. [#5784](https://github.com/GetStream/stream-chat-android/pull/5784)
 - Introduce `ChatClient.uploadFile`, `ChatClient.deleteFile`, `ChatClient.uploadImage`, and `ChatClient.deleteImage`, to upload/delete files/images that are not related to any channel. [#5909](https://github.com/GetStream/stream-chat-android/pull/5909)
+- Add `NewMessageEvent.channelMessageCount` field representing the total number of messages in the channel. [#5918](https://github.com/GetStream/stream-chat-android/pull/5918)
+- Add `MessageDeletedEvent.channelMessageCount` field representing the total number of messages in the channel. [#5918](https://github.com/GetStream/stream-chat-android/pull/5918)
+- Add `channelRole` property on the `Message` model, representing the role of the member who sent the message in the channel. [#5919](https://github.com/GetStream/stream-chat-android/pull/5919)
 
 ### ⚠️ Changed
 
@@ -45,6 +52,7 @@
 ### ⬆️ Improved
 
 ### ✅ Added
+- Add `ChannelState.messageCount` field representing the total number of messages in the channel. [#5918](https://github.com/GetStream/stream-chat-android/pull/5918)
 
 ### ⚠️ Changed
 
@@ -80,8 +88,11 @@
 - Filter out Poll Options differing only on whitespaces. [#5913](https://github.com/GetStream/stream-chat-android/pull/5913)
 
 ### ⬆️ Improved
+- Improved accessibility of the message list items. [#5911](https://github.com/GetStream/stream-chat-android/pull/5911)
+- Show quoted message over messages with attachments. [#5924](https://github.com/GetStream/stream-chat-android/pull/5924)
 
 ### ✅ Added
+- Added `MessageReactionList` and `MessageReactionItem` component factories. [#5925](https://github.com/GetStream/stream-chat-android/pull/5925)
 
 ### ⚠️ Changed